Create RSS feeds
This change creates the proper feeds for events and upcoming CFP’s.

It also updates the blog feed properly and sets the head elements to find the proper feeds.

Note that the location of the “speaking” page needs to change to support this. An issue will be opened on devopsdays-web for this.

Signed-off-by: Matt Stratton <matt.stratton@gmail.com>