[refs #00611] Add custom_url to program
Add ability to add a custom_url type for all types of talk on the
program page. Adding custom_url: "https://valid.url.co.uk" will result
in the link on the schedule being opened in another window.

I also took it upon myself to change the indenting in the single.html
file but this can be reverted.