updates CFP OPEN footer to use time comparions
The existing footer takes the current time (build time), casts it to a
string, and compares it with the CFP start/end times which are also
being re-cast to strings here (twice!).

In order to allow us to use timezones as part of our comparison and not
have the formatting involved in comparisons, use the `time` template
function and directly compare against `now` without typecasting.

ref https://gohugo.io/functions/time/
ref https://gohugo.io/functions/now/