Add barebones of tito shortcode logic
I want to be able to add a shortcode for tito and add specific named
params in order to display the tito widget on a page I wish.

I also want to allow use of releases and discount-codes

The current naming convention is using underscores so lets do that!
We'll also move to using external css and js, rather than keep them
locally we don't want to add extra work always updating these.

Signed-off-by: Chris Mills <me@christophermills.co.uk>